Fix arm then immediate disarm then arm again loop.
Found issue with immediate disarm, caused by the blocking nature of the arm sequence which would make the last update of the MPU be too long causing an immediate disarm. The MPU would then update and everything would be set to start arming causing a loop. Made the arm function non-blocking by adding a state and calling it in the drone loop.